Admin users can delete  and create default users

+++ PoliceSystem/DAL/UserDaoImpl.cs
@@ -26,6 +26,7 @@ namespace PoliceSystem.DAL
                 else
                 {
                     //laat de gebruiker weten dat de username bezet is
+                    throw new InvalidOperationException("Username already exists");
                 }
             }
         }
@@ -70,22 +71,58 @@ namespace PoliceSystem.DAL
             }
         }
 
-        public void Remove(User user)
+        public List<User> getAllUsers()
         {
-            if (UserExists(user))
+            using (PoliceDbContext db =new PoliceDbContext())
             {
-                using (PoliceDbContext db = new PoliceDbContext())
-                {
-                    db.Users.Remove(user);
-                }
-            }
-            else
-            {
-                //User doesnt exists.
+                return db.Users.Include(u => u.UserGroup).ToList();
             }
         }
 
+        public void Remove(int id)
+        {
+            using (PoliceDbContext db = new PoliceDbContext())
+            {
+                User user = FindById(id);
+                if (UserExists(user))
+                {
+                    if (user.UserGroup.Name.Equals("admin"))
+                    {
+                        throw new InvalidOperationException("Cannot delete admin account");
+                    }
+
+                    db.Users.Attach(user);
+                    db.Users.Remove(user);
+                    db.SaveChanges();
+                }
+
+                else
+                {
+                    //User doesnt exists.
+                    throw new InvalidOperationException("User with username: " + user.Username + " does not exist");
+                }
+            }
+        }
+
+        /// <summary>
+        /// Checks if the user exists in the database, based on the combination of the id and username
+        /// </summary>
+        /// <param name="user">The user with username and id</param>
+        /// <returns>True if the user exists, false if not</returns>
         public bool UserExists(User user)
+        {
+            using (PoliceDbContext db = new PoliceDbContext())
+            {
+                return db.Users.Any(u => u.Username == user.Username && u.Id == user.Id);
+            }
+        }
+
+        /// <summary>
+        /// Check if the user credentials are valid
+        /// </summary>
+        /// <param name="user">The user with username and password</param>
+        /// <returns>True if valid, false if not valid</returns>
+        public bool IsValid(User user)
         {
             using (PoliceDbContext db = new PoliceDbContext())
             {
